fix requesting too big array.

+++ src/main/java/com/hankcs/algorithm/AhoCorasickDoubleArrayTrie.java
@@ -939,7 +939,7 @@ public class AhoCorasickDoubleArrayTrie<V> implements Serializable
                 {
                     // progress can be zero // 防止progress产生除零错误
                     double toSize = Math.max(1.05, 1.0 * keySize / (progress + 1)) * allocSize;
-                    double maxSize = Integer.MAX_VALUE * 0.95;
+                    int maxSize = (int) (Integer.MAX_VALUE * 0.95);
                     if (allocSize >= maxSize) throw new RuntimeException("Double array trie is too big.");
                     else resize((int) Math.min(toSize, maxSize));
                 }
